javadoc comment + small tweek odoCorrection

+++ src/ca/mcgill/ecse211/localization/UltrasonicLocalizer.java
@@ -147,7 +147,7 @@ public class UltrasonicLocalizer {
     
     // Stop rotating
     this.driver.stop();
-    float angle = (enteringAngle + exitingAngle)/2.0;
+    double angle = (enteringAngle + exitingAngle)/2.0;
     return angle;
   }
   
@@ -193,7 +193,7 @@ public class UltrasonicLocalizer {
     
     // Stop rotating
     this.driver.stop();
-    float angle = (enteringAngle + exitingAngle)/2.0;
+    double angle = (enteringAngle + exitingAngle)/2.0;
     return angle;
   }
